+++ pandas/__init__.py
@@ -44,11 +44,7 @@ from pandas.core.sparse.api import *
 from pandas.stats.api import *
 from pandas.tseries.api import *
 from pandas.core.computation.api import *
-
-from pandas.tools.concat import concat
-from pandas.tools.merge import (merge, ordered_merge,
-                                merge_ordered, merge_asof)
-from pandas.tools.pivot import pivot_table, crosstab
+from pandas.core.reshape.api import *
 
 # deprecate tools.plotting, plot_params and scatter_matrix on the top namespace
 import pandas.tools.plotting
@@ -58,9 +54,7 @@ scatter_matrix = pandas.util.decorators.deprecate(
     'pandas.scatter_matrix', pandas.plotting.scatter_matrix,
     'pandas.plotting.scatter_matrix')
 
-from pandas.tools.tile import cut, qcut
-from pandas.tools.util import to_numeric
-from pandas.core.reshape import melt
+from pandas.core.reshape.util import to_numeric
 from pandas.util.print_versions import show_versions
 from pandas.io.api import *
 from pandas.util._tester import test
