+++ pandas/core/computation/expr.py
@@ -13,7 +13,7 @@ from pandas.compat import StringIO, lmap, zip, reduce, string_types
 from pandas.core.base import StringMixin
 from pandas.core import common as com
 import pandas.io.formats.printing as printing
-from pandas.tools.util import compose
+from pandas.core.reshape.util import compose
 from pandas.core.computation.ops import (
     _cmp_ops_syms, _bool_ops_syms,
     _arith_ops_syms, _unary_ops_syms, is_term)
