+++ pandas/core/reshape/util.py
