+++ pandas/io/formats/format.py
@@ -168,7 +168,7 @@ class SeriesFormatter(object):
         self._chk_truncate()
 
     def _chk_truncate(self):
-        from pandas.tools.concat import concat
+        from pandas.core.reshape.concat import concat
         max_rows = self.max_rows
         truncate_v = max_rows and (len(self.series) > max_rows)
         series = self.series
@@ -409,7 +409,7 @@ class DataFrameFormatter(TableFormatter):
         Checks whether the frame should be truncated. If so, slices
         the frame up.
         """
-        from pandas.tools.concat import concat
+        from pandas.core.reshape.concat import concat
 
         # Column of which first element is used to determine width of a dot col
         self.tr_size_col = -1
