+++ pandas/tests/reshape/data/allow_exact_matches.csv
