+++ pandas/tests/reshape/test_hashing.py
