+++ pandas/tests/reshape/test_join.py
@@ -12,7 +12,7 @@ from pandas import DataFrame, MultiIndex, Series, Index, merge, concat
 
 from pandas._libs import join as libjoin
 import pandas.util.testing as tm
-from pandas.tests.tools.test_merge import get_test_data, N, NGROUPS
+from pandas.tests.reshape.test_merge import get_test_data, N, NGROUPS
 
 
 a_ = np.array
