+++ pandas/tests/reshape/test_merge.py
@@ -9,8 +9,8 @@ import random
 
 import pandas as pd
 from pandas.compat import lrange, lzip
-from pandas.tools.concat import concat
-from pandas.tools.merge import merge, MergeError
+from pandas.core.reshape.concat import concat
+from pandas.core.reshape.merge import merge, MergeError
 from pandas.util.testing import assert_frame_equal, assert_series_equal
 from pandas.core.dtypes.dtypes import CategoricalDtype
 from pandas.core.dtypes.common import is_categorical_dtype, is_object_dtype
