+++ pandas/tests/reshape/test_merge_ordered.py
