+++ pandas/tests/reshape/test_union_categoricals.py
