+++ setup.py
@@ -642,6 +642,7 @@ setup(name=DISTNAME,
                 'pandas.core.dtypes',
                 'pandas.core.indexes',
                 'pandas.core.computation',
+                'pandas.core.reshape',
                 'pandas.core.sparse',
                 'pandas.errors',
                 'pandas.io',
@@ -673,7 +674,6 @@ setup(name=DISTNAME,
                 'pandas.tests.series',
                 'pandas.tests.scalar',
                 'pandas.tests.tseries',
-                'pandas.tests.tools',
                 'pandas.tests.plotting',
                 'pandas.tools',
                 'pandas.tseries',
@@ -703,7 +703,7 @@ setup(name=DISTNAME,
                                         'data/html_encoding/*.html',
                                         'json/data/*.json'],
                     'pandas.tests.io.formats': ['data/*.csv'],
-                    'pandas.tests.tools': ['data/*.csv'],
+                    'pandas.tests.reshape': ['data/*.csv'],
                     'pandas.tests.tseries': ['data/*.pickle'],
                     'pandas.io.formats': ['templates/*.tpl']
                     },
